Add support to parse parquet files
This commit adds feature to accept parquet file as
1) target index dataset file 2) neighbor dataset
3) query dataset . It also supports partition.
For every partition, one iterator will be created.
Since, parquet file can be very big and it doesn't
support index based loading, we will use batch iterator
and load one batch at a time and extract vectors for
given start and end offset.
We will also use variable to hold batch if complete
batch is not processed. Once the batch is processed,
it will clear the variable.

Signed-off-by: Vijayan Balasubramanian <balasvij@amazon.com>